// src/types/language.ts
export enum Language {
  ARABIC = 'ar',
  AZERBAIJANI = 'az',
  BELARUSIAN = 'be',
  BULGARIAN = 'bg',
  BENGALI = 'bn',
  BOSNIAN = 'bs',
  CATALAN = 'ca',
  CZECH = 'cs',
  DANISH = 'da',
  GERMAN = 'de',
  GREEK = 'el',
  ENGLISH = 'en',
  ESPERANTO = 'eo',
  SPANISH = 'es',
  ESTONIAN = 'et',
  FINNISH = 'fi',
  FRENCH = 'fr',
  HEBREW = 'he',
  HINDI = 'hi',
  CROATIAN = 'hr',
  HUNGARIAN = 'hu',
  INDONESIAN = 'id',
  ICELANDIC = 'is',
  ITALIAN = 'it',
  JAPANESE = 'ja',
  GEORGIAN = 'ka',
  KANNADA = 'kn',
  KOREAN = 'ko',
  CORNISH = 'kw',
  LATVIAN = 'lv',
  MALAYAM = 'ml',
  MARATHI = 'mr',
  NORWEGIAN = 'no',
  NORWEGIAN_BOKMAL = 'nb',
  DUTCH = 'nl',
  PUNJABI = 'pa',
  POLISH = 'pl',
  PORTUGUESE = 'pt',
  ROMANIAN = 'ro',
  RUSSIAN = 'ru',
  SLOVAK = 'sk',
  SLOVENIAN = 'sl',
  SERBIAN = 'sr',
  SWEDISH = 'sv',
  TAMIL = 'ta',
  TELUGU = 'te',
  TETUM = 'tet',
  TURKISH = 'tr',
  UKRAINIAN = 'uk',
  URDU = 'ur',
  IGPAY_ATINLAY = 'x-pig-latin',
  SIMPLIFIED_CHINESE = 'zh',
  TRADITIONAL_CHINESE = 'zh-tw'
}